Click header to go back
This was difficult and may be considered to be incorrect:
* Browse -> Corpus -> Home -> Back := Browse
* Add a key to 'home' to allow it to be refreshed
* Detect a navigation from 'Home', and replace history in that case
* Remove link coloring when clicked for the title

Also convert components to allow for hooks

Now: only allow one page on the back stack

In 2021, Max asked that 'back' would navigate to the main page if on a document.
The problem here is that navigating to a document pushes history
But the browser has no ability to 'pop' the history.

To fix this:
* on the first navigation, 'push'
* always go 'back'